refactor: extract ActivateSkillTool into tools/activate_skill.rs
Reduce tools/mod.rs from ~900 to ~850 lines by moving the
ActivateSkillTool struct and its Tool trait impl into a dedicated file.
